parche

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/expedienteComercial/ExpedienteComercialManager.java
@@ -4501,7 +4501,7 @@ public class ExpedienteComercialManager extends BusinessOperationOverrider<Exped
 				} else {
 					TmpClienteGDPR tmpClienteGDPR = genericDao.get(TmpClienteGDPR.class,
 							genericDao.createFilter(FilterType.EQUALS, "numDocumento", dto.getNumDocumento()));
-					if(!Checks.esNulo(tmpClienteGDPR.getIdAdjunto())) {
+					if(tmpClienteGDPR != null && !Checks.esNulo(tmpClienteGDPR.getIdAdjunto())) {
 						docAdjunto = genericDao.get(AdjuntoComprador.class,
 								genericDao.createFilter(FilterType.EQUALS, "id", tmpClienteGDPR.getIdAdjunto()));
 					}
@@ -4542,7 +4542,9 @@ public class ExpedienteComercialManager extends BusinessOperationOverrider<Exped
 				ofertaApi.resetPBC(expediente, true);
 				
 				TmpClienteGDPR tmpClienteGDPR = genericDao.get(TmpClienteGDPR.class, genericDao.createFilter(FilterType.EQUALS, "numDocumento", comprador.getDocumento()));
-				clienteComercialDao.deleteTmpClienteByDocumento(tmpClienteGDPR.getNumDocumento());
+				if(tmpClienteGDPR != null){
+					clienteComercialDao.deleteTmpClienteByDocumento(tmpClienteGDPR.getNumDocumento());
+				}
 
 				return true;
